refactor indices code into small pieces
